requirements.txt
google-cloud-firestore==2.11.0
requests==2.31.0
Flask==2.3.3
